+++ tests/Main.hs
@@ -3,6 +3,7 @@ import qualified Util as T
 import qualified CanonicalizePath
 import qualified CopyFile001
 import qualified CopyFile002
+import qualified CopyFileWithMetadata
 import qualified CreateDirectory001
 import qualified CreateDirectoryIfMissing001
 import qualified CurrentDirectory001
@@ -25,6 +26,7 @@ main = T.testMain $ \ _t -> do
   T.isolatedRun _t "CanonicalizePath" CanonicalizePath.main
   T.isolatedRun _t "CopyFile001" CopyFile001.main
   T.isolatedRun _t "CopyFile002" CopyFile002.main
+  T.isolatedRun _t "CopyFileWithMetadata" CopyFileWithMetadata.main
   T.isolatedRun _t "CreateDirectory001" CreateDirectory001.main
   T.isolatedRun _t "CreateDirectoryIfMissing001" CreateDirectoryIfMissing001.main
   T.isolatedRun _t "CurrentDirectory001" CurrentDirectory001.main